fixes plugin executor import cycle and executor selection
When a plugin is made with a custom Operator and executor, an import cycle occurs when the executor is chosen in airflow.cfg because the executors/__init__.py starts loading plugins too early when a non standard executor is chosen.
changed DEFAULT_EXECUTOR use to a function call which returns the default executor. This lazy approach fixes the import cycle.

revision eb5982d4aac135051666d2977bbf6adfc8b9e2a7 (included in 1.8) breaks plugin executors altogether. It makes a new module for every plugin, so import statements need to be adapted, but the executor selection is left unchanged, so it ends up assigning the plugin module as an executor.
fixed executor selection to work with the new plugin modules system. in Airflow.cfg a executor can now be specified as {pugin_name}.{executor_name}

Fixes:
 - https://issues.apache.org/jira/browse/AIRFLOW-860