[AIRFLOW-1150] Fix scripts execution in sparksql hook[]
When using the the SparkSqlOperator and submitting
a file (ending with
`.sql` or `.hql`), a whitespace need to be
appended, otherwise a Jinja
error will be raised.

However the trailing whitespace confused the hook
as those files will
not end with `.sql` and `.hql`, but with `.sql `
and `.hql `. This PR
fixes this.

In the test, I've added the `get_after` function
to easily check if the
path is really stripped or not by the `-f` option.

Closes #2259 from gglanzani/master